requirements.txt
slack-bolt>=1.18.0
python-dotenv>=0.19.0
aiohttp>=3.8.0